added functions from XeApp code to load ventilation data files
I just took the LoadReadData functions and removed the pieces that were not needed for this project (e.g. loading diffusion/gas exchange data or loading DICOM Philips images). The function is more concise now and thus easier to debug. Loading data into the workspace where it can be processed is a crucial starting point for the code.